backend/pyproject.toml
[project]
name = "sentiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.13",
    "bs4>=0.0.2",
    "fastapi[standard]>=0.115.11",
    "pydantic-settings>=2.8.1",
    "pytest>=8.3.5",
    "sortedcontainers>=2.4.0",
    "spacy>=3.8.4",
]
